pyproject.toml
[project]
name = "analytics-command-center"
version = "0.1.0"
description = "Ops Analytics Command Center for NYC Taxi trips"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dbt-core==1.7.0",
    "dbt-postgres==1.7.0",
    "dagster>=1.7.0",
    "dagster-webserver>=1.7.0",
    "dagster-dbt>=0.23.0",
    "dagster-postgres>=0.23.0",
    "great_expectations==0.18.8",
    "pandas==2.2.0",
    "pyarrow==15.0.0",
    "fastparquet==2024.2.0",
    "sqlalchemy==2.0.25",
    "psycopg2-binary==2.9.9",
    "requests==2.31.0",
    "plotly==5.18.0",
    "scipy==1.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts", "alerts", "orchestration"]

[tool.dagster]
module_name = "orchestration.definitions"
